Finish program
I added the remaining forthe main function in parse_build_halfedge_off
and kept the edges dictionary in HalfedgeMesh. Note I might remove
get_halfedge in the future--it seems unneeded due to self.edges.
Simplified __hash__ for Halfedge because when I kepted all the member
variables as part of the hash they popped the stack with an infinite
loop.